[OPIK-872] Unify Create and Edit prompt modals.
