+++ apps/opik-frontend/src/components/pages/PromptPage/PromptTab/PromptTab.tsx
@@ -9,7 +9,7 @@ import CodeHighlighter, {
   SUPPORTED_LANGUAGE,
 } from "@/components/shared/CodeHighlighter/CodeHighlighter";
 import UseThisPromptDialog from "@/components/pages/PromptPage/PromptTab/UseThisPromptDialog";
-import EditPromptDialog from "@/components/pages/PromptPage/PromptTab/EditPromptDialog";
+import EditPromptVersionDialog from "@/components/pages/PromptPage/PromptTab/EditPromptVersionDialog";
 import CommitHistory from "@/components/pages/PromptPage/PromptTab/CommitHistory";
 import usePromptVersionsById from "@/api/prompts/usePromptVersionsById";
 import usePromptVersionById from "@/api/prompts/usePromptVersionById";
@@ -141,7 +141,7 @@ const PromptTab = ({ prompt }: PromptTabInterface) => {
         promptName={prompt.name}
       />
 
-      <EditPromptDialog
+      <EditPromptVersionDialog
         key={editPromptResetKeyRef.current}
         open={openEditPrompt}
         setOpen={handleOpenEditPrompt}
